Merge pull request !101 from bean/feat-autotest

+++ bos/b_type.h
@@ -127,7 +127,7 @@ extern "C" {
 
 #define B_UNUSED(x) (void)(x)
 
-#define B_ALIGN_TO_2_BYTES(addr) (((uint32_t)(addr) + 1) & ~((uint32_t)1))
+#define B_ALIGN_TO_2_BYTES(addr) ((((uintptr_t)(addr)) + 1) & ~((uintptr_t)1))
 
 /**
  * \}
